# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "utopia API",
        "description": "API documentation for utopia ",
        "contact": {
            "email": "info@utopia.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://gestion.utopia.madrid/api/",
            "description": "utopia API"
        }
    ],
    "paths": {
...
        "/categories": {
            "get": {
                "tags": [
                    "categorias"
                ],
                "summary": "Mostrar categorias",
                "operationId": "31d7eb1fb499fb981765e01717f25c7c",
                "responses": {
                    "200": {
                        "description": "Respuesta Ok."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            },
            "post": {
                "tags": [
                    "categorias"
                ],
                "summary": "Crear categoria",
                "operationId": "9d672c849a554dcdeb0c9328926d679e",
                "requestBody": {
                    "description": "Post categoria",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "idsector"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "_"
                                    },
                                    "idsector": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "1"
                                    },
                                    "informacion": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "_"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Respuesta Ok."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
        "/categories/{id}": {
            "get": {
                "tags": [
                    "categorias"
                ],
                "summary": "Mostrar categoria ",
                "operationId": "3c38bf6848eab35a9980f2152312a3f1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Parámetro necesario para la consulta de categorias",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "int": {
                                "summary": "Introduce un número de id .",
                                "value": "1"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Respuesta Ok."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            },
            "put": {
                "tags": [
                    "categorias"
                ],
                "summary": "Modificar categoria ",
                "operationId": "94463130a9fb896f028c0e1d05589408",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Parámetro para editar categorias",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "int": {
                                "summary": "Introduce un número de id .",
                                "value": "1"
                            }
                        }
                    }
                ],
                "requestBody": {
                    "description": "put categoria",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "idsector"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "_"
                                    },
                                    "idsector": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "1"
                                    },
                                    "informacion": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "_"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Respuesta Ok."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            },
            "delete": {
                "tags": [
                    "categorias"
                ],
                "summary": "Eliminar categoria",
                "operationId": "917ad5212d304fa6c5221f39e73189aa",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Parámetro para editar categorias",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "int": {
                                "summary": "Introduce un número de id .",
                                "value": "1"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Respuesta Ok."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
...
    }
}